dd-9 Move adapter and dialog to mainScreen package
