+++ app/src/main/java/io/krugosvet/dailydish/android/mainScreen/MealListAdapter.kt
@@ -1,4 +1,4 @@
-package io.krugosvet.dailydish.android
+package io.krugosvet.dailydish.android.mainScreen
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
+import io.krugosvet.dailydish.android.R
 import io.krugosvet.dailydish.android.db.objects.Meal
 import io.krugosvet.dailydish.android.utils.getFormattedDate
 import io.krugosvet.dailydish.android.utils.getMeals
