+++ app/src/main/java/io/krugosvet/dailydish/android/mainScreen/MealListPageFragment.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import io.krugosvet.dailydish.android.MealListAdapter
 import io.krugosvet.dailydish.android.R
 import io.krugosvet.dailydish.android.db.objects.Meal
 import io.krugosvet.dailydish.android.utils.RealmFragment
